Regression(r196563): It is no longer possible to call window.addEventListener without an explicit 'this' https://bugs.webkit.org/show_bug.cgi?id=154245
Reviewed by Ryosuke Niwa.

Source/WebCore:

This patch adds support for calling the EventListener API without an
explicit 'this' value. If no explicit 'this' value is passed, then we
fall back to using the global object. This matches Chrome and Firefox's
behavior. It also fixes the Dromaeo/cssquery-dojo.html test.

Test: fast/dom/Window/addEventListener-implicit-this.html

* bindings/scripts/CodeGeneratorJS.pm:
(GenerateFunctionCastedThis):

LayoutTests:

Add a layout test to cover the use of the EventListener API without an
explicit 'this' value.

* fast/dom/Window/addEventListener-implicit-this-expected.txt: Added.
* fast/dom/Window/addEventListener-implicit-this.html: Added.

Identifier: 172359@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@196588 268f45cc-cd09-0410-ab3c-d52691b4dbfc