Add an ensure function on HashMap that takes a key and a function to make the lazy value initialization idiom easier https://bugs.webkit.org/show_bug.cgi?id=134857
Reviewed by Geoffrey Garen.

Source/WTF:

The current idiom for adding a value to a HashMap if the key is not already present, to allow for not
unnecessarily constructing the new value if it won't be used, is:

{
    auto result = map.add(key, Value());
    if (!result.isNewEntry)
        return result.iterator->value;

    result.iterator->value = createNewValue();
    return result.iterator->value;
}

or sometimes

{
    auto& value = map.add(key, Value()).iterator->value;
    if (!value)
        value = createNewValue();
    return value;
}

With this patch, you can now use the new function HashMap::ensure(key, functor). This will allow us to
change to using the following idiom:

{
    return map.ensure(key, [] { return createNewValue(); });
}

The passed in functor will only be called if the key is not already present in the HashMap.

* wtf/HashMap.h:
(WTF::HashMapTranslator::translate):
(WTF::HashMapEnsureTranslator::hash):
(WTF::HashMapEnsureTranslator::equal):
(WTF::HashMapEnsureTranslator::translate):
(WTF::HashMapTranslatorAdapter::hash):

Tools:

* TestWebKitAPI/Tests/WTF/HashMap.cpp:
Add tests for HashMap::ensure.



Identifier: 172470@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@196716 268f45cc-cd09-0410-ab3c-d52691b4dbfc