Web Inspector: Text Align segmented control blinks while editing other properties in Visual Styles sidebar https://bugs.webkit.org/show_bug.cgi?id=154487 <rdar://problem/24754703>
Patch by Devin Rousso <dcrousso+webkit@gmail.com> on 2016-02-20
Reviewed by Timothy Hatcher.

The icon list property editor blinking issue was caused by the fact that
the selected value was toggled on/off each time the value was set on the
editor. In order to prevent this, the logic for the setter value() was
modified to just match a keyword icon to the given value and select it.

* UserInterface/Views/VisualStyleKeywordIconList.js:
(WebInspector.VisualStyleKeywordIconList.prototype.set value):
(WebInspector.VisualStyleKeywordIconList.prototype._handleKeywordChanged):

Identifier: 172605@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@196863 268f45cc-cd09-0410-ab3c-d52691b4dbfc