REGRESSION(r195949): [GTK] Test /webkit2/WebKitWebView/insert/link is failing since r195949 https://bugs.webkit.org/show_bug.cgi?id=153747
Reviewed by Michael Catanzaro.

Source/WebCore:

Do not return early when reaching a boundary if there's a range
selection. In that case, the selection will be cleared and
accessibility will be notified.

Test: editing/selection/move-to-line-boundary-clear-selection.html

* editing/FrameSelection.cpp:
(WebCore::FrameSelection::modify):

LayoutTests:

Add test to check that moving to line boundary clears the
selection even if the cursor is already at the boundary.

* editing/selection/move-to-line-boundary-clear-selection-expected.txt: Added.
* editing/selection/move-to-line-boundary-clear-selection.html: Added.

Identifier: 172730@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@197024 268f45cc-cd09-0410-ab3c-d52691b4dbfc