<g> wrapping <symbol> causes display of hidden <symbol> https://bugs.webkit.org/show_bug.cgi?id=154576
Patch by Said Abou-Hallawa <sabouhallawa@apple.com> on 2016-02-26
Reviewed by Darin Adler.
Source/WebCore:

The SVGSymbolElement is allowed in the shadow tree of an SVGUseElement
only if it subtree root element. Any descendant SVGSymbolElement should
be removed from the subtree because it is a hidden container. If the cloned
subtree includes an SVGUseElement which references an SVGSymbolElement,
the same rule will be applied to the descendant SVGUseElement. The goal
is to remove all the descendant SVGSymbolElements from the cloned target
because these SVGSymbolElements will be expanded to SVGSVGElements and
hence become visible.

* svg/SVGUseElement.cpp:
(WebCore::disassociateAndRemoveClones): A helper function which removes
cloned SVGElements and their subtrees from their parents and disassociate
them from their originals.

(WebCore::removeDisallowedElementsFromSubtree): Use disassociateAndRemoveClones().

(WebCore::removeSymbolElementsFromSubtree): Removes all the descendant
SVGSymbolElements from the cloned subtree. It does not remove the root
element itself if it is an SVGSymbolElement because this one will be
expanded to an SVGSVGElement which is exactly what we need.

(WebCore::SVGUseElement::cloneTarget): Call removeSymbolElementsFromSubtree()
to remove the descendant SVGSymbolElements from the cloned subtree before
appending it to the container shadow root.

LayoutTests:

Ensure the <symbol> element is not displayed when it's wrapped in a <g>
element and this <g> element is referenced by a <use> element.

* platform/gtk/svg/custom/use-on-g-containing-symbol-expected.png: Removed.
* platform/gtk/svg/custom/use-on-g-containing-symbol-expected.txt: Removed.
* platform/ios-simulator/svg/custom/use-on-g-containing-symbol-expected.txt: Removed.
* platform/mac/svg/custom/use-on-g-containing-symbol-expected.png: Removed.
* platform/mac/svg/custom/use-on-g-containing-symbol-expected.txt: Removed.
* platform/win/svg/custom/use-on-g-containing-symbol-expected.txt: Removed.
* svg/custom/use-on-g-containing-symbol-expected.svg: Added.
* svg/custom/use-on-g-containing-symbol.svg:
The original test was wrong. It had the following definition:
    "<g id='symbol'><symbol>...</symbol></g>"
And it was expecting to have the symbol drawn if the <g> element was
referenced like that
    "<use xlink:href='#symbol'/>"
FireFox does not render anything for this <use> element which is correct.
With this patch, this test failed so it had to be modified to test the right
behavior. Also it is now converted to a ref test.

Identifier: 172837@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@197194 268f45cc-cd09-0410-ab3c-d52691b4dbfc